requirements.txt
boto3==1.21.8
botocore==1.24.8
imageio==2.14.1
matplotlib==3.5.1
nltk==3.7
numpy
Pillow==9.0.1
requests==2.27.0
tensorly==0.7.0
torch==1.10.2
torchvision==0.11.3
